Cleaned Up CSS
Removed `#bg` tag since it is referencing nothing in the `html`

Set the `#bgimg` position to absolute so that it does not adjust itself
the other elements on the page

Set other things such as `hero-content` to `position: relative` so they
can adjust to each other. Removed margin adjustments as they were being
used to force elements into moving and when position was changed to
relative they were causing issues.

Added a class called `communication` to the other <section> tags on the
html and set all communication to be in the middle of the page

Removed text shadow on sign up button because it hurt my eyes.