More Changes
Moved communication div out of hero content section and put it with the
other communication divs so it can be adjusted accordingly.

There was a a left move on the sign up in the hero content that was
skewing the content so I took it out to centre it.

Moved stuff around so that its in the right places. Used pixels for this